% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utilities.R
\name{attachPgDb}
\alias{attachPgDb}
\title{Attach a postres database to a duckdb database}
\usage{
attachPgDb(con, uri)
}
\arguments{
\item{con}{DuckDb connection}

\item{uri}{URI for the postgres database}
}
\value{
DuckDb connection with postgres database attached
}
\description{
Attach a postres database to a duckdb database
}
