% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkOmopOnPostgres.R
\name{benchmarkOmopOnPostgres}
\alias{benchmarkOmopOnPostgres}
\title{Benchmark OMOP on Postgres}
\usage{
benchmarkOmopOnPostgres(cdm, n_iterations = 1)
}
\arguments{
\item{cdm}{CDM reference}

\item{n_iterations}{Number of iterations to run}
}
\value{
Benchmarking results
}
\description{
Benchmark OMOP on Postgres
}
