% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utilities.R
\name{duckdbHasPgDbAttached}
\alias{duckdbHasPgDbAttached}
\title{Check if a PostgreSQL database is attached to DuckDB}
\usage{
duckdbHasPgDbAttached(con)
}
\arguments{
\item{con}{DuckDb connection}
}
\value{
TRUE if postgres database attached, otherwise FALSE
}
\description{
Check if a PostgreSQL database is attached to DuckDB
}
