% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utilities.R
\name{duckdbIsUsingPgDb}
\alias{duckdbIsUsingPgDb}
\title{Check if duckdb is using Postgres as default catalog}
\usage{
duckdbIsUsingPgDb(con)
}
\arguments{
\item{con}{DuckDb connection}
}
\value{
TRUE if using Postgres as default catalog, otherwise FALSE
}
\description{
Check if duckdb is using Postgres as default catalog
}
