% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utilities.R
\name{duckdbUseDuckDb}
\alias{duckdbUseDuckDb}
\title{Ensure duckdb is using the native DuckDB catalog}
\usage{
duckdbUseDuckDb(con)
}
\arguments{
\item{con}{DuckDb connection}
}
\value{
DuckDb connection that uses native DuckDB as default catalog
}
\description{
Ensure duckdb is using the native DuckDB catalog
}
