% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utilities.R
\name{duckdbUsePgDb}
\alias{duckdbUsePgDb}
\title{Ensure duckdb is using Postgres as default catalog}
\usage{
duckdbUsePgDb(con)
}
\arguments{
\item{con}{DuckDb connection}
}
\value{
DuckDb connection that uses Postgres as default catalog
}
\description{
Ensure duckdb is using Postgres as default catalog
}
