% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{getUserTables}
\alias{getUserTables}
\title{Get tables created by a user in a schema.}
\usage{
getUserTables(src, schema = "public", user = NULL)
}
\arguments{
\item{src}{It can either be a cdm_reference, a postgres_source or a
DBI connection object.}

\item{schema}{Character. The schema to search in.}

\item{user}{Character. Users to filter by. If NULL no filter is applied.}
}
\value{
Tibble with the identified tables.
}
\description{
Get tables created by a user in a schema.
}
