% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresSource.R
\name{localPostgres}
\alias{localPostgres}
\title{This function creates a connection to the local postgres instance}
\usage{
localPostgres(client = Sys.getenv("TEST_PG_DRIVER", "RPostgres"))
}
\arguments{
\item{client}{Client. Can be "RPostgres", "adbc", "odbc", or "DatabaseConnector".}
}
\value{
A connection to your postgres local instance
}
\description{
It relays on environmental variables such as:
\itemize{
\item dbname = \code{OMOP_POSTGRES_CONNECTOR_DB}
\item host = \code{OMOP_POSTGRES_CONNECTOR_HOST}
\item port = \code{OMOP_POSTGRES_CONNECTOR_PORT}
\item user = \code{OMOP_POSTGRES_CONNECTOR_USER}
\item password = \code{OMOP_POSTGRES_CONNECTOR_PASSWORD}
}
}
\examples{
\dontrun{
library(OmopOnPostgres)

localPostgres()
}
}
