% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockPostgresCdmReference.R
\name{mockPostgresCdmReference}
\alias{mockPostgresCdmReference}
\title{Mock Postgres CDM reference}
\usage{
mockPostgresCdmReference(client, datasetName = "GiBleed", cdmVersion = NULL)
}
\arguments{
\item{client}{Client supported by OmopOnPostgres}

\item{datasetName}{Mock dataset to include. See
'omock::availableMockDatasets()' for available
datasets.}

\item{cdmVersion}{CDM version, see
'omock::mockCdmFromDataset()' for more details.}
}
\value{
CDM reference for mock data in a local
Postgres database
}
\description{
Mock Postgres CDM reference
}
