% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopPostgresConnector-package.R
\docType{package}
\name{OmopOnPostgres-package}
\alias{OmopOnPostgres}
\alias{OmopOnPostgres-package}
\title{OmopOnPostgres: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OmopOnPostgres}
  \item Report bugs at \url{https://github.com/ohdsi/OmopOnPostgres/issues/}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
