% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R, R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{expectedIndexes}
\alias{existingIndexes}
\alias{statusIndexes}
\alias{createIndexes}
\alias{createTableIndex}
\alias{listSourceTables}
\alias{insertTable}
\alias{dropSourceTable}
\alias{insertCdmTo}
\alias{readSourceTable}
\alias{cdmDisconnect}
\alias{cdmTableFromSource}
\alias{compute}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:compute]{compute()}}}

  \item{omopgenerics}{\code{\link[omopgenerics:cdmDisconnect]{cdmDisconnect()}}, \code{\link[omopgenerics:cdmTableFromSource]{cdmTableFromSource()}}, \code{\link[omopgenerics:createIndexes]{createIndexes()}}, \code{\link[omopgenerics:createTableIndex]{createTableIndex()}}, \code{\link[omopgenerics:dropSourceTable]{dropSourceTable()}}, \code{\link[omopgenerics:existingIndexes]{existingIndexes()}}, \code{\link[omopgenerics:expectedIndexes]{expectedIndexes()}}, \code{\link[omopgenerics:insertCdmTo]{insertCdmTo()}}, \code{\link[omopgenerics:insertTable]{insertTable()}}, \code{\link[omopgenerics:listSourceTables]{listSourceTables()}}, \code{\link[omopgenerics:readSourceTable]{readSourceTable()}}, \code{\link[omopgenerics:statusIndexes]{statusIndexes()}}}
}}

