% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R, R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{expectedIndexes}
\alias{existingIndexes}
\alias{statusIndexes}
\alias{createIndexes}
\alias{createTableIndex}
\alias{listSourceTables}
\alias{insertTable}
\alias{dropSourceTable}
\alias{insertCdmTo}
\alias{readSourceTable}
\alias{cdmDisconnect}
\alias{cdmTableFromSource}
\alias{compute}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{compute}}}

  \item{omopgenerics}{\code{\link[omopgenerics]{cdmDisconnect}}, \code{\link[omopgenerics]{cdmTableFromSource}}, \code{\link[omopgenerics]{createIndexes}}, \code{\link[omopgenerics]{createTableIndex}}, \code{\link[omopgenerics]{dropSourceTable}}, \code{\link[omopgenerics]{existingIndexes}}, \code{\link[omopgenerics]{expectedIndexes}}, \code{\link[omopgenerics]{insertCdmTo}}, \code{\link[omopgenerics]{insertTable}}, \code{\link[omopgenerics]{listSourceTables}}, \code{\link[omopgenerics]{readSourceTable}}, \code{\link[omopgenerics]{statusIndexes}}}
}}

